Use partial class and methods to separate netfx and core
